RONDB-113: Failure to allocate more than 2 Gbyte of memory in one allocation
The alloc_impl function did always report Out of memory when requesting to
allocate more than 2 GByte of memory in one go. This can happen when using
automatic memory allocation and the total memory exceeds something on the
order of 25 GByte. Thus a fairly serious bug that requires a new release of
RonDB 22.01.

Also removed a faulty require check.