HOPSWORKS-2922: Step 36
In the new scheme of things we only update extent position based on
in which category it fits into. This means that we have to send
in the before and after bits into update_extent_pos from
the callback of disk_page_prealloc. The disk_page_prealloc decreased
the level before retrieving the page from disk. But this calculation
was very conservative and will often need adjustment when the page
comes into memory.