HOPSWORKS-2922: Step 39
shirnk_tuple can now be called also without any variable sized memory
part where a var part is forced. This means that setting sizes[MM]=1
lead to allocation of a variable sized part which is lost from memory
and eventually we crash on an assert that checks that the variable
sized parts are not more than the fixed size parts. Fixed by setting
it to 0, which previously also was done in expand_tuple.