RONDB-114: Missing sync on initialisation of files
When initialising REDO logs, Tablespace files and UNDO log files we use
the flag OM_INIT that will invoke writes in a loop of the pages in the file.
We prepare this by setting the autosync rate to 1 MByte. However we
will never check this by calling sync_on_write. Thus if the file system cannot
keep up with the write speed, we can use massive amounts of memory in
Linux leading to Out Of Memory situations.

Fixed by adding a call to sync_on_write in the write loop.
Also increased the 1 MByte to 16 MByte.