HOPSWORKS-2922: Step 44, fix issue with INSERT-UPDATE
After INSERT followed by UPDATE the row could change size and this could
lead to that we need to change the disk page used. However to know this
we need to know the free and used variables of the disk page. To
ensure this we need to set req_struct->m_disk_page_ptr.p also, not only
i-value also in this particular case.