HOPSWORKS-2922: Step 50
When writing log records which are divided into two parts we needed to
be more careful in calculating the extra overhead to the free log space.
Updated print_file.cpp to handle the new UNDO log record types.

Added proper handling of variable sized disk rows for nr_delete which is used
to delete records during Copy Fragment, this was previously missed in the
implementation.