HOPSWORKS-2922: Step 66
With variable sized disk rows we don't know for certain
which extent index it will end up in disk_page_alloc.
This depends on what happens to the directory structure
if it stays at the same size or if it increases. This
means that the extent index could change for variable
sized disk rows in disk_page_alloc.

disk_page_free already handled this in a correct manner
for both variable sized disk rows and fixed size disk rows.