HOPSWORKS-2922: Step 67
Major changes of handling of row size changes of disk rows.

One major insight was that we cannot return any allocated
row parts unless we moved to another page that can house the
largest row size we have seen so far during the transaction.
The problem is if we abort one or more operations, in this case
we could potentially return to a size which we haven't got space
for. Thus moving to a new disk page we have to allocate at least
as much as the largest size we have seen so far.

If we have to increase the size of a row without changing page
we will remember this extension even if later we resize the row
to become smaller. Thus if we need to allocate a new page we
have already seized the largest size available. Thus if we need
a new page the size must be the largest so far.

To handle this we need some special handling of partial aborts.
Both to handle the case with new disk pages and with resized
original disk page.

One more insight is that since we have to work with 2 disk pages
in the state DISK_REORG, this means that we have to ensure that
during reading in those two pages that we don't page out the first
page while reading in the second page.

To handle this we added the flags REF_REQ and DEREF_REQ in get_page
towards PGMAN. These calls increase the busy_count same as already
happening at COMMIT_REQ. COMMIT_REQ is releasing the pages when
update_lsn is called for the page.

For prepare phase we use a special call with only DEREF_REQ set
to release the page such that it can be paged out in the next
signal.

We needed to ensure that the proper increment and decrement is
used when calling update_extent_pos, this is especially true for
calls to disk_page_alloc that previously didn't change the
extent position which it now can do.

Fairly major rewrite of handle_size_change_after_update that was
reorganised to make it easier to understand the flow of the code.

Removed m_max_disk_part_size since the deletes also read in the
disk pages at commit time and prepare time.