HOPSWORKS-2922: Step 68
Mistakes around m_uncommitted_used_space on operation record.
It is important to move this to the next operation in a multi-row
operation. Thus need copying to next operation in prepareActiveOp.
Also when committing it is important to decrease uncommitted_used_space
on page with m_uncommitted_used_space from operation record. This does
not happen by inserting the row into the variable disk page.