HOPSWORKS-2922: Step 73
With variable sized disk rows we need access to the disk page choosen
when performing UPDATEs and re-inserts on rows that started with an
initial insert operation. This means that we need to set the disk
row reference in the new row and not only in the copy row in this
case. This disk page needs to be read into memory before starting
the operation.

When extending the size of a disk row on a new page we need to
also set the allocated space in the key.m_page_idx used to store
this information temporarily during the transaction.

Added a bit more jam's for debugging.