HOPSWORKS-2922: Step 74
When creating a new table we need to check that all nodes support
the version where variable sized disk data is introduced. We should
however avoid skipping support simply because a node is down while
creating the table.

When altering the table it is important to transport the setting
of variable sized disk data support in Dbdict. Similarly such support
must exist in NDB API.