HOPSWORKS-2922: Step 77
m_use_varsized_disk_data is only prepared for debugging, so will not
use it in checks of equalness. It is set by the data nodes dependent
on which versions are running in the cluster.

Found a bug in the new ndb malloc library around allocation of the smallest
size (16 bytes) and combining this with allocation a memory area, releasing
it and allocating it again. A new test case was created for this as well.

In addition the 16 bytes areas was put in list 0, but we never allocated
from this area. Minor check added for this in malloc routine.