RONDB-124: Concurrent release of same page possible in Suma
When we enter into an out of buffer release process we can still
simultaneously get a FIRE_TRIG_ORD that discovers also that
some pages need to be released. This should be avoided when
out of buffer release is ongoing since that will release all
pages anyways.

Fixed by introducing a state variable whether out of buffer
release is ongoing or not.