RONDB-126: Fix for neighbours in send threads
An issue discovered before is that moving from neighbour to non-neighbour
transporters can lose signals. This was fixed for the case when we activate
multi transporters.

However every time we call the triplet of
startChangeNeighbourNode
setNeighbourNode
endChangeNeighbourNode

We start this by disabling all neighbour nodes. After that we add one or
more nodes as neighbour nodes. However in QMGR we only add back the
node we are activating or deactivating. This means that each time we
call this triplet any neighbour nodes that we don't handle will
move from neighbour node to non-neighbour node. Thus all transporters
of those nodes could potentially lose some signal. This could lead to
that e.g. TCGETOPSIZECONF is not sent and the GCI protocol could
stop.

The fix is to only remove the node from the neighbour list that we
are dealing with. The other neighbour transporters will remain there.

There was code in DBDIH to move around which nodes should use
neighbour nodes. However this code could lead to lost signals as well.
The list of potential neighbour nodes is decided at startup and not
changed after that now. It is not a big issue if we don't reorganise
for formation of added node groups. The nodes will still be able to
communicate in an efficient manner.