RONDB-126: New fix for neighbours in send threads
The previous fix was unsuccessful and this patch reverts that patch
and implements a new fix.

The fix is built on top of the previous fix of neighbour nodes
where we introduced the function insert_activate_trp to insert
transporters that have been put in list as neighbours and now
need to be inserted in list of transporters for non-neighbours.

The code disables all neighbour transporters and moves them to
non-neighbour status in startChangeNeighbourNodes. The change
is that startChangeNeighbourNodes calls insert_activate_trp for
all transporters moved to non-neighbour status.

We also call flush_send_buffers to ensure that any signals sent
is visible to sender.

This patch doesn't solve the problem that some nodes in the same
node group no longer are treated as neighbours. This is handled
in a later patch.