HOPSWORKS-2922: Step 78
The DISK_REORG bit is set in the committed tuple header. The reason is that we
need to recall that we have allocated an extra page. Fixed a few places where we
still looked up the DISK_REORG bits in the copy row header bits.

The variable m_load_extra_diskpage_on_commit is set to indicate that we need to
read in the extra disk page during commits. This variable was never set, it only
needs to be set when we set the DISK_REORG flag. It needs to be copied over to
the next operation.

Also when an operation that set DISK_REORG and m_load_extra_diskpage_on_commmit is
aborted, we need to ensure that the flags are moved to the new last operation
since the allocation of the new page part must not be forgotten.

An assert in commit_operation asserted that the diskPagePtr is equal to the
disk row key found in the copy row. This is not true for operations that set
DISK_REORG so moved those asserts.