RONDB-130: Fixes for O_DIRECT
O_DIRECT is activated using the method set_direct_io.
This method assumes that O_DIRECT will work with 512 byte
alignment and offsets. If not setting O_DIRECT will fail.

Failing O_DIRECT is mostly ok, but it isn't ok for Tablespace
files. In this case the code assumes that a write is also
sync:ed which it will be if O_DIRECT is set. But if not set
we need to set the O_SYNC flag to ensure that each write is
a sync as well. If this flag doesn't work we have to issue
a fsync on each write call. This was a clear bug on Mac OS X
where O_DIRECT isn't supported.

O_DIRECT_SYNC_FLAG should not be supported with sparse REDO
log initialisation.

Moved the setting of O_DIRECT to after initialising the files.
The reason is that we cannot discover if O_DIRECT works on an
empty file or a new file. It will always fail.

A consequence of this patch is that new LCP and backup files
will never use O_DIRECT since the set_direct_io call will always
fail on new files.