RONDB-131: Fixes of LQH_TRANSREQ protocol
When more than 2 alive nodes participate in Node failure handling
we handle 2 at a time to avoid overloading the TC handling the
node failure. However we made a mistake in using the node id of
the node completing this phase as the failed node id.

In addition we didn't respect the limit of at most 2 nodes at a time
when starting new nodes.