Bug#34312769 FIXES ATTRIBUTE COMPARISON LENGTH/SIZE
Bug caused the assert:
NdbSqlUtil.cpp:498: require((lb + m1 <= n1 && lb + m2 <= n2)) failed

When a VARCHAR column is part of the primary key and is being
attempted to be updated, incorrect length of the 's2' argument,
which is the value read from the database, was supplied to the cmp_attr() methode.
That triggered the above require, which checked the consistency of
the supplied arguments with the length in the AttrInfo headers.

Patch also fixed an incorrect 'length' check which require the
binary byte length of the arguments to be the same: That is not
true for attributes being compared as characters, as different compare
semantics for different character/collation may compare binary different
strings as equal - as defined by each character set / collations.

-> Moved that check into the code sections where we binary compare
the attributes.

Change-Id: I47de0fab181f6a59a1aa7ac6ee5a8b8dd1d056a5