RONDB-150: Fix of a mutex issue
MySQL 8.0.30 introduced a variable lockCount on each fragment in DBACC.
This was updated every time a lock was taken and released. Previous
to RonDB 22.01 this would only happen in LDM threads. However with the
changes in RonDB 22.01 and now in RonDB 22.10 it is now the case that
also query threads can update this variable.

However the mutex is split into 4 parts, thus it is necessary to also
split this variable into 4 parts, each part protected by its own mutex.

Eventually this variable will be moved away from production code. It is
to costly to maintain and causes a fair amount of cache line jumping
between CPU cores. It should be sufficient to have it activated in all
test runs.