RONDB-152: New handling of pollReceive/performReceive
This patch is based on some observations using KUtrace.
It was noted that in the NDB API it was most of the time
sufficient to fetch from a single transporter in
performReceive.

One call to performReceive could easily lead to many
wakeups of other threads. It is important to avoid
waiting to wakeup other threads. Thus it is important
to immediately start handling those when discovered.

A potential problem this could lead to is that we
spend too much time calling epoll_wait and we could
also easily get unfairness. To avoid this we introduce
an iterative approach in pollReceive.

This means that as long as there is still work to do
in performReceive from the previous pollReceive we
will simply return immediately from pollReceive.
When we set up a new iteration we set up two bitmaps.

Each transporter has a socket, when poll/epoll indicates
that there is data on this socket we set a bit in the
m_recv_socket_transporters bitmap.

On TCP transporters we will also set a bit in the bitmap
of m_read_transporters. For SHM transporters we will set
m_read_transporters dependent on if there is data in
the SHM buffer.

During one iteration we will handle things differently
in the data node and in the NDB API. In the data node we
are very often working with very short executions, this
means that every microsecond extra waiting will lead to
increased latency. Thus in this case we will always only
read from transporter at a time to ensure quick handling
of incoming signals.

For the NDB API we've noted that this approach is a bit
too eager and instead we check the number of bytes
received, if a message is larger than 2 kBytes we will
exit the performReceive function. Also we will at most
read from 4 transporters.

During one iteration we will still handle a maximum number
of signals. This means that we could have more data in the
transporter buffers after an iteration. This is indicated
in the m_has_data_transporters bitmap. This bitmap is
reset at the beginning of each iteration. Each transporter
that has more data will then set a bit during the iteration
in this bitmap.

At start of a new iteration all bits set in m_has_data_transporters
will be set in m_read_transporters (not in m_recv_socket_transporters
since we have no indication of any more data on socket, unless
epoll_wait saw it).

Now there is one more issue, this is job buffer full conditions.
In this case we will stop at a certain transporter and only
receive from sockets. This means that at the next iteration we
need to ignore unpacking from the transporters before the
stop transporter id. We do however need to receive on the socket
for all transporters. Thus we reset all bits before the stop
transporter id when the stop transporter id was set (which only
happens when we see a job buffer full condition).

Thus for those transporters we might read on the socket, but we
will not unpack any signals.