RONDB-155: Reorganise query thread handling
Reorganise the query thread handling. Previously query threads
were separate threads executing on the same CPU core as the
LDM thread. This worked well, but has some drawbacks for write
intensive loads. For write-heavy loads the query threads were
hardly used. In hyperthreaded CPUs this had not such a great
impact. However in e.g. ARM CPUs that only have 1 CPU per core
it means that full cores goes without usage when locking threads
to CPUs.

In this patch we change the query threads to be a part of the LDM
threads. The query blocks are included in the LDM threads. Thus
LDM threads can both execute as LDM thread and as a query thread.
Thus each LDM thread can be used for reads of all other LDM threads.

We still have recover threads that only execute as query threads
during restart.

We have changed the scheduler such that all LDM threads use the
same scheduling mechanism as previously used by LDM threads.
We prefer to use the LDM thread owning the data, next we prefer
the LDM thread executing on the same CPU core. Finally we use
a load balanced scenario on all threads sharing L3 cache.

To find the LDM thread using the same CPU core required a bit of
code. If no same CPU on the CPU core exists or if we are using
ThreadConfig we will ignore checking for another LDM thread
before using the load balancing loop.