RONDB-157: More eager sending
Analysing RonDB using kutrace which provides detailed information
about what goes on down to a granularity of 10 nanoseconds showed
that we could be more eager to send messages over the network at
times.

The NDB API is waiting for messages from the data node, however not
all messages are going to wake a thread in the NDB API. Thus we went
through the code and analysed all actions that will lead to a wake
up of a thread in the NDB API. These actions could potentially lead
to sending eagerly. However sending eagerly only interacts with send
threads. We don't want to use block threads for sending as long as
they have to work to do. However if they are not occupied it is fair
that they could assist with sending.

Thus at each time we find reason to send we will first flush all
signals and then we will attempt to wake up the send thread if it
isn't already awake.

RonDB data nodes can become very batch-oriented when many threads are
sending data to the node at the same time. This could lead to a ping-pong
effect with the application. The application is busy sending messages to
RonDB, next RonDB is busy executing the messages. But there is free CPU
resources in RonDB while APIs are busy and vice versa.

There is no need to consider flushing when signals are sent inside the
data node since this is already handled by the send signal routines
ensuring that signals sent to local threads will not be delayed more
than necessary.

It is however important to not count all signals, we only want to
keep track of those signals that actually wake up some waiting
thread or thread of execution at the other node. Thus we increment
m_send_wakeups when sending the following signals to other nodes:

Scan messages
-------------
SCAN_FRAGCONF
SCAN_FRAGREQ
SCAN_FRAGREF
SCAN_TABCONF

Transaction messages
--------------------
COMMIT
COMMITTED
COMPLETE
COMPLETED
ABORT
ABORTED
ABORTREQ
ABORTCONF

API interaction for transactions
--------------------------------
TCKEYCONF
TC_COMMITCONF
TCROLLBACKCONF

Interactions with DBTC for key lookups
--------------------------------------
LQHKEYCONF
LQHKEYREF

A special one is simple dirty reads from applications. This is handled
already in execution of LQHKEYREQ for simplicity to avoid complex
new code in TRANSID_AI.

There is no need to handle all other signals, they are not real-time
critical in the same manner as the normal key lookups and scan
signals.

Removed the flush assistance, always assist send if assist at all.
Added assist of send thread when reaching another boundary based on
number of wakeups.

Removed m_max_signals_before_send* and added m_max_send_wakeups_before_assist
and tied them to the SchedulerResponsiveness.

Adjusted numbers of SchedulerResponsiveness

Before going to sleep it is a good idea to assist the send thread.
We have checked that there is no other action to do, thus if we
go to sleep we will be sleeping for a while anyways, thus it is
a good idea to assist the send thread before we fall asleep.

Also if we decided to assist the send thread we should perform
this action even if the send thread isn't awake. However if
we call it from handle_scheduling_decisions we will only assist
if the send thread is already busy, we probably still have other
important work to perform, so no reason to assist a sleeping
send thread in this case.

We will never assist with sending to more than one transporter
per do_send call.

Increase minimum number of wakeups from receive thread. This ensures
that we won't leave to the wakeup thread too quickly.

Assist send thread during spin every 25 microseconds in block threads,
but not in receive thread.