RONDB-160: Mix Query threads and non-Query threads caused crash
The variable recBlockNo is setup with the value V_QUERY to indicate
that we should use a query block when executing the query. This
variable was set up using the primary replica nodes setting on
query threads. If the node that actually was selected to execute
the query didn't have query threads, but the primary replica node
had, this would cause a crash in the node actually executing the
query.

Added a new test case ndb_basic_mix_numcpus with two data nodes,
one using NumCPUs=2 (no query thread) and one node using
NumCPUs=4 (using query thread) to ensure that the bug is tracked
in the future.