RONDB-167: ClusterJ supporting setting database when retrieving Session object
ClusterJ has been limited to handle only one database per cluster connection.
This severely limits the usability of ClusterJ in cases where there are many
databases such as in a multi-tenant use case for RonDB.

At the same time the common case is to handle only one database per cluster
connection. Thus it is important to maintain the performance characteristics
for this case.

One new addition to the public ClusterJ API is the addition of a new getSession
call with a String object representing the database name to be used by the
Session object. Once a Session object has been created it cannot change to
another database. A session object can have a cache of DTO objects, this would
be fairly useless when used with many different databases. Thus this isn't
supported in this implementation. The limitation this brings about is that a
transaction is bound to a specific database.

We can cache sessions in RonDB, we have one linked list of cached session objects
for the default database. Other databases create a linked list at first use of
a database in a SessionFactory object. The limit on the amount of cached Session
objects is maintained globally. Currently we simply avoid putting it back on
the list if the maximum has been reached. An improvement could be to have a
global order of the latest use of a Session object, this hasn't been implemented
here.

A Session object has a Db object that represents the Ndb object used by the
Session. This Ndb object is bound to a specific database. For simplicity we
store the database name and a boolean if the database is the default database.
The database name could have been retrieved from the Ndb object as well.
This database name in the Db object is used when retrieving an NdbRecord for the
table.

ClusterJ handles NdbRecord in an optimised manner that tries to reuse them as
much as possible. Previously it created on Ndb object together with an
NdbDictionary object to handle NdbRecord creation. Now this dictionary is renamed
and used for only the default database. Each new database will create one more
Ndb object together with an NdbDictionary object. This object will handle all
NdbRecord's for that database. For quick finding of this object we use a
ConcurrentHashMap using database name to find this NdbDictionary object.

Previously there was a ConcurrentHashMap for all NdbRecord's, both for tables
and for indexes. These used a naming scheme that was tableName only or
tableName+indexName.

This map is kept, but now the naming scheme is either databaseName+tableName or
databaseName+tableName+indexName.

Thus more entries are likely to be in the hash map, but it should not affect
performance very much.

These maps are used to iterate over when unload schemas and when removing cached
tables.