Bug #23539857 - WL#8069: "LCP_SCANNED_BIT == 0 FAILED" does not hold.
Partial LCP needs each fragment LCP to be able to determine the
precise state of the fragment at the start of the LCP and the precise
diff of the fragment between the start of this LCP and the start of
the last LCP. This is tracked using row header information and page
header information.  In cases where physical pages are removed this is
also tracked in logical page-map information.

Assume a page existed and is included in the current LCP.  Before the
LCP-scan reaches it, it is released due to commit/rollback of some
operation on the fragment releasing the last used storage on the page.

Since the released page cannot be found by the scan, the release
itself set the LCP_SCANNED_BIT of the page-map entry it was mapped
into, in order to indicate that the page is already 'handled' from the
point of view of the current LCP.  Consecutive alloc and release of
the pages mapped to the entry during the Lcp are ignored.  The state
of the entry at the start of the LCP was also set as 'Allocated' in
the page map entry.

Problem: These settings are cleared only when the next LCP is
prepared. Any page release associated to the page-map entry before the
clearance would violate the offended requirement that the bit is not
set.

Fix: Remove the incorrect requirement.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>

Change-Id: I04a5691a959a474e95287c2a943824a9efb5e7ee