HOPSWORKS-2922: Step 86
Introduced a new get_page request to pgman.
The request is called DIRTY_HEADER. Its intended use is to flag to pgman that
the page is dirty since the page header has been updated. However the page isn't
necessarily dirty such that the local checkpoint needs to care about it.

This will set a new flag on the page entry D_HEADER to ensure that we don't
cleanup a page in this state until it has been paged out to disk.

For simplicity we clear this flag as soon as the state DIRTY is set and we
avoid setting the D_HEADER state if DIRTY is already set.