HOPSWORKS-2922: Step 87
When a size change of a disk row occurs during an UPDATE/INSERT/WRITE
operation and the size change makes the row bigger, we need to preallocate
space on the page. This is accomplished by updating information on the
page itself. However pages doing UPDATE in prepare phase has not set the
page as dirty. This means that the page can be cleaned up without being
written to disk. This will lose the preallocation information and will
lead to a crash.

To handle this we introduce a new method disk_page_dirty_header that
ensures that the page is using the new DIRTY_HEADER interface in PGMAN
to ensure that the page header updates aren't lost when the page is
cleansed, the page must be written to disk before being cleansed.

A number of new comments are added around this.

Also ensuring that the 0x8000 bit in list_index is always set when the
page is on disk. Otherwise we could come into a situation where there
is no consistency between this variable and the presence in the dirty
page list in DBTUP.