RONDB-155: Fixes related Query thread used for Locked Reads
When query threads are colocated with LDM threads we must handle
ABORT handling and fix the code to handle this.

Had to handle checks for a signal had already executed in a
fashion that takes into account that signal id can wrap
after a while.

ABORT signal can have up to 7 words in it for those cases
to carry information required to find the LDM thread and
to carry the thread id and signal id of the signal that
sends the ABORT signal to the query worker.