HOPSWORKS-2922: Step 89, fix for DISK_REORG state
During COMMIT of a change to a disk row that entered the state
DISK_REORG we need to release the old disk row. However the
reference to this old disk row is overwritten early on in
the commit processing as part of copying the fixed memory part
from copy row to memory row. Thus we need to copy out the
disk reference before we start the commit processing into a
stack variable that can be used further down in the commit
processing when handling the DISK_REORG state.