RONDB-179: unlink of pid file bug
The deletion of the pid file used a pointer to an area that
was already free'd, this led to a deletion of some random
thing, most of the time unlink failed since the file name
was simply an empty string. However in rare cases with
usage of systemd one could delete some real file.