RONDB-190: Fix issue with reorg of varpage
Prepare Insert and CommittedRead can happen concurrently,
Prepare Insert in an LDM thread and CommittedRead in a Query
thread. This is normally ok since Prepare Insert doesn't
actually update the row memory until commit. However it does
allocate memory.

The problem is that allocation of row memory can sometimes cause
a page for variable sized data to be reorganised. This cannot happen
concurrently with a reader that reads that particular page. In
very rare circumstances this could lead to a crash.

Fixed by ensuring that those reorg calls are only done when holding
exclusive access to the fragment.