RONDB-190: Reversed allocation order of fixed and varsized
When INSERT is performed we need a varsize row and a fixed size row.
The varsize row could require an upgrade to exclusive access due
to compaction of varsize page. The fixed size row requires a mutex
on the TUP page to ensure that the row isn't seen by scans before
it is properly setup.

Previously the order was to first allocate the fixed row and next
the varsize row. This led to deadlock when the varsize row had to
use exclusive access. So by reversing the order we are safe from
this issue.