RONDB-190: We mustn't upgrade to exclusive lock when executing metadata operations
During metadata operations we allocate space for default values on the
default fragment record. This allocation happens in LDM thread always
and requires no upgrade to exclusive access since the default fragment isn't
used for normal operations and also the fragment isn't created yet, so not
accessible for user operations.