RONDB-199: Ensured that pid file contains PID of data node, not of angel
The data node uses two processes, an angel process which is the first
process started, the angel is daemonized after which it is forked into
another process which is the real data node process.

When interacting with environments like systemd it is easier to handle
this if the pid file contains the pid file of the real data node process.

Stopping the angel process doesn't stop the data node process. Thus
keeping track of this PID isn't of any great value.

Thus in using RonDB data nodes in combination it is recommended to not
set StopOnError to 1 since this means that the angel will restart the
ndbmtd no matter how it stopped. Thus it is better to set StopOnError=0
and to use the pid file (ndb_NODEID.pid in the same directory as the
log files) to find the data node pid to stop or kill.