RONDB-215: Handle deprecation of connect-retries and connect-delay
The command options connect-retries and connect-delay to ndbmtd have been deprecated
since there are similar options for generic NDB programs. However to ensure that scripts
that use this feature will continue to function we reinsert those command options, set
the default of them to 0. This means that if they are not set they will be ignored.
If they are set they will be used.