RONDB-217: Handling of all-recv threads had issues
When running with 0 LDM threads and all LDM instances in recv threads
we forgot to set recv_threads to 0 in calculation of shared instances
and CPU locking (recv threads are setup as if they were LDM threads).
There was also a flaw in the calculation of the shared instance of
the LDM threads.