RONDB-225: More partitions for special tables
The default is that a table uses 2 partitions per data node in the RonDB cluster.
This works quite well when the cluster spreads writes and reads to different
tables.

In a 2-node cluster this means that the data nodes will use 4 partitions
per table. For writes this means that at most 4 CPUs will work on updates
for this table. This limits throughput on that table to about 400k updates
per second. So not so common that this will be a bottleneck.

For reads we have query thread capability that will ensure that up to a
minimum of 16 CPUs can work on the table and even 32 CPUs in some HW setups.
This means that the table will be able to handle 2-3 million reads per
second.

However there are still applications where this could potentially be a problem.
As an example the DBT2 benchmark has one table that stands for 75% of the
CPU usage in the RonDB cluster. Since DBT2 is very write heavy, this means
that this table will easily become a bottleneck.

To handle this a number of new variants of table comment have been added.

This table creation:
create table t(id int NOT NULL PRIMARY KEY, data char(8))
engine=ndb
comment="NDB_TABLE=PARTITION_BALANCE=FOR_RP_BY_LDM_X_2";

It will create 2 * PartitionsPerNode partitions per data node.

We have also addded:
FOR_RP_BY_LDM_X_4 => 4 * PartitionsPerNode per data node
FOR_RP_BY_LDM_X_6 => 6 * PartitionsPerNode per data node
FOR_RP_BY_LDM_X_8 => 8 * PartitionsPerNode per data node
FOR_RP_BY_LDM_X_16 => 16 * PartitionsPerNode per data node

Thus the partitioning is still dependent on the configuration
setting of PartitionsPerNode which defaults to 2.

Increasing the PartitionsPerNode increases the number of partitions
to scan when performing range scans. Thus for tables that sustain
heavy scan activity one need to be careful in increasing the number
of partitions in the table.